moved to use isolated server for aic
