+++ packages/aic/src/doc/signature.ts
